dded support for the filtering by component.
Added support for filtering by regular expression.
Finished up with the support of the options.
Save the appender's type in the appender.
Added support to know whether an appender can be duplicated or not (i.e. syslog has to be unique, file can be any number.)
Added support for the "path" field in the file appender.
Fixed a bug in the appender::set_config() where I would reset the f_severity field on every call, which is wrong.
Marked the components and configuration done.
Made the `--debug` also for the addition of a console appender.
Updated the README.md a bit. It still needs a lot of help.
Various clean ups.