Added a dependency on CppThread so we can use it for our mutex, gettid() and asynchronous login.
Added a few variables now that we have access to the option environment.
Changed the DEFINE_LOGGER_VARIABLE() to directly be followed by a function body.
Added a private_logger to encompass nearly all the globals so the destruction of the logger along the thread works as expected.
Changed all the factories accordingly, especially, we do not use global shared pionters, not even for the g_instance variable.
Prevent the re-allocation of the g_instance variable (it would not work well.)
Added a lowest severity calculation in the logger so we can avoid wasting time when an incoming message is going to be thrown away anyway.
Fixed the environment get_...() functions, all the strings are returned by copy now. Added a guard too.
Fixed the variable param class too, strings are returned by copy and it has guards.
Fixed the SNAPLOGGER_VAR_FUNCTION string spelling, the "i" was missing.
Fixed many #include.