Added a fatal error exception so we can throw when a fatal error gets logged. By default this feature is off. Use the set_fatal_severity() to use the feature.
Added the '<' operator to the appender class to sort by severity.
Applied a fix to the send_message() test of components, if empty consider it's "normal".
Fixed the buffer_appender::process_message(), the formatted_message already includes a newline.
Fixed the buffer_appender derivation so the stringstream is public.
Added a lowest severity variable in the logger so we can immediately turn off the generation of a message.
Fixed the reset(), the lowest severity needs to be set to OFF again.
When adding an appender we need to verify the severity level, it could be lower than existing appender.
Fixed the severity_changed() function, if the new level is higher, go through the list of appenders to define the lowest severity.
Properly check whether a message severity level is OFF, if so, always replace the buffer with our null_buffer.
Added a "<<" operator to automatically print the name of a severity level.
Added a message.cpp test to test that the severity levels work as expected (and a few other things).