Implemented the advgetopt log callback to capture those errors.
Changed the implementation of the options check in regard to severity. It does the same but it is cleaner.
Tweaked the system severity aliases and descriptions.
Made the add_alias() reactive so if the severity is already registered, the alias automatically gets added.
Made the get_description() return the name if the string is empty.
Removed the get_aliases(), it was a _duplicate_ of get_all_names() and was not implemented.
Added many more unit tests (appender & severity and also in existing files).
Fixed the default f_options_files_directory, it has to be /etc/snaplogger.
Enhanced the README.md to be closer to what is now implemented.
Removed the has_severities(), it was not used.