Implemented the default severity support.
Use the DEFAULT_SEVERITY in two places, instead of a hard coded default.
Fixed the severity.ini, we need to use `severity=...` not `level=...`.
Added the `verbose` severity to the severity.ini.
Fixed the severity levels of those that were pushed by `verbose`.
Added a SEVERITY.md and README.md in the /usr/share/snaplogger/... folders.
Use advgetopt::is_true() instead of `... == "true"`.
Testing without the `O_NOCTTY` when opening the console.