Added the `--logger-show-banner` and `--logger-hide-banner` options.
Allow tools to hide the logger banner by default.
Added a "banner" component.
Added the `--list-severities` command.
Changed the process_logger_options() to continue on errors so all the command line options get processed and the caller can decide whether to continue or not by testing the result.
Added a test against process_logger_options() in the tools and stop immediately on errors.
Added some documentation.