SNAP-742: Added the `systemd` format for the `severity` variable.
Made the severity to syslog priorty a separate static function.
Fixed tests to work with the newer advgetopt version.
Added a test to verify the severity to syslog tag.
Tweaked test to verify severity number is output with "format=number".
Tested all our severity levels in the Print Severity test.