Added two new severity levels.
Added the SEVERE severity level which is higher than ERROR but lower
than CRITICAL.

Added the CONFIGURATION severity level, which can be used to output logs
about configuration parameters. This is lower a little lower than the
INFORMATION severity so in most cases, by default you do not get that
output at all.