Fixed bug testing tellp().
When the priority is lower (using the temporary use 'X' as the lowest priority
feature), it generates a buffer which tellp() function returns -1. In that
case we completely avoid outputting the message. Before it would only avoid
outputting the message if tellp() returned exactly 0.

Also made a small update to support the new serverplugins. i.e. pass the
server identifier to the server.