Added support for a tty=... parameter.
The console appender prints out data whether the console is a TTY or not.
Now, we can turn off that printing if the console is not a TTY by setting
the tty=true parameter.

Really added the new postinst and sysuser files so we will get the snaplogger
user and the log directories.

Updated the changelog accordingly.