Added scripts to help create/destroy log folders.
Each project uses a sub-directory under /var/log/<project>/... to save their
log files. The new scripts help to create and destroy those properly with
a secure set of permissions.

The destroy part uses the shredlog tool in various manners to make sure that
the deletion happens appropriately. To make that work appropriately, I
upgraded the shredlog tool (use snapdev instead of duplicated code or libboost).

Moved the is_hdd() to the library, although it still doesn't work on
DigitalOcean. Created a corresponding tool.

Add a default 50-shredlog.conf file so we can put the selected mode=... so
we don't force the default to "delete" (which is the safest at the moment).