Replaced some throw with a return.
When adding a parameter which includes the "${" sequence to a log, it is
viewed as a format introducer. This fix is to avoid throwing when the format
looks invalid. Instead we just return an error and move on. There are still
some throw instruction at lower levels. Things that should never occur.