Use std::source_location for location.
Using std::source_location makes it a lot easier and better as it includes
a column and it solves a few issues we were having with __FILE__, __LINE__,
etc. Added a column field and functions to get/set that value. And also
changed the line/column type to std::uint_least32_t like what the standard
library uses.

Added missing severity definitions (which I did not do when I added a
few earlier).

Fixed the output of severity min/max (STRINGIZE does not work in this case).
Also added tests for duplicated severity, and the minimum and maximum.

Fixed some documentation.