Adding ANSI converter.
I often want to send console output to a website because it makes it
easier to review the data (for build systems type of a thing) and
whenever the process generates ANSI code, especially to change colors
it looks like crap. This converter will instead transform the ANSI
codes in HTML tags as required (i.e. <b> for bold, a <span> with a
style attribute to change colors, etc.)