Added support for fallback appenders.
The file and syslog appenders had some hard coded fallbacks (file supports
a console or syslog and the syslog can also write messages to a console).

Now, we can defined the fallback_appenders=... parameter with a list of
appender names. That list gets used if sending a message through one
appender fails.