Fixed two command line options.
The --logger-file and --syslog-identity would not work with the newer version
of advgetopt because I did not add the "dynamic configuration" flag. That
meant the set_value() did not do anything and silently skipped setting
anything.

The test on whether an appender was already used as a fallback was not
properly detected and instead nothing was being logged.

Updated some docs to make certain things clearer.