Fixed another fallback bool issue.
Prevent auto-creation of console logger if the user also used one of the
--log-file, --log-config, --syslog, --no-log, or --console (although the
last one would not be duplicated anyway).

This is done because in general if you use --trace or --debug, then you
probably want to see the logs in your console. However, if you do
--log-file, then you probably want those logs in that file and not
in the console.