Updated to the newer version of serverplugins.
The old version would do a strange job at creating the "server" object.
It still included a strong concept of instance(). That was removed.
One of our object has to hold on a serverplugins::collection object
which has all the plugins and the server attached to. No more instances
required at all (although, internally, there is a singleton which handles
the loading of each library--the repository).

This is much cleaner and will clearly separate the snaplogger plugins
from any other server plugins.